moved VideoScreen from __init__ to own file
